Add option to add a label to a custom field
This feature should be helpful for determining builds that integrated code associated with a given JIRA issue and, conversely, finding JIRA issues that were integrated into a given build. Unlike Fix Versions, the label content does not need to be pre-defined. Issue-level labels were unsuitable in our case as issue-level labels were already used for another purpose.

As labels cannot contain spaces, this addition treats spaces as a delimiter. (The space delimiter piece was tested against a live JIRA instance.) As JIRA appears to treat label fields as a set, no deduplication is needed.

There does not appear to be a way to query a custom field type and operations on non-label custom field may silently fail.